Removed redundant cast.

+++ mule-shell-server/test-java/net/sf/zoftwhere/mule/resource/AccountResourceTest.java
@@ -343,7 +343,7 @@ class AccountResourceTest extends TestResource<AccountResource> {
 		}).orElse(0L);
 
 		assertNotNull(count);
-		assertEquals((long) roleModelArray.length, (long) count, "");
+		assertEquals(roleModelArray.length, (long) count, "");
 	}
 
 	private void registerTestAccountGroup() {
